Add option to TeamCity hook to just check for pending changes
(needs review)

The default behaviour of 'Always running a build' (with the 'add2Queue'
parameter) is not always desirable, because some TeamCity setups can
have carefully configured rules about when to /actually/ do a build
which are ignored by this direct approach. This commit adds the option
to just poke the TeamCity instance to check for changes (effectively the
same as hitting the 'Check for pending changes' option in the TeamCity
web user interface), which allows any TC build-rules to be applied.

The 'checkForChangesBuildType' parameter is not well documented, but it
is directly referenced by a JetBrains employee Yegor Yarko here:

http://youtrack.jetbrains.com/issue/TW-4134#comment=27-9255